# public/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "dev-ocean-api",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "http://172.18.0.1:86/api/"
        }
    ],
    "paths": {
...
        "/v1/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Create a new user.",
                "operationId": "3932d2c8bad547ad03e75d30a7bca311",
                "requestBody": {
                    "$ref": "#/components/requestBodies/UserRegisterRequest"
                },
                "responses": {
                    "201": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MiddlewareError"
                                },
                                "example": {
                                    "message": "You are logged in already."
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors response.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Validation"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/v1/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Enter to your account.",
                "operationId": "268f05a6d83a13e1da762571fea969be",
                "requestBody": {
                    "$ref": "#/components/requestBodies/LoginRequest"
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/User"
                                        },
                                        "Bearer": {
                                            "$ref": "#/components/schemas/Bearer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error: Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MiddlewareError"
                                },
                                "example": {
                                    "message": "Credentials do not match."
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MiddlewareError"
                                },
                                "example": {
                                    "message": "You are logged in already."
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors response.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Validation"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/v1/profile": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Shows your profile.",
                "operationId": "ebdd19f6563ee215399f6e23933fef8c",
                "responses": {
                    "200": {
                        "description": "Shows user profile",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Error: Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Unauthenticated."
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/v1/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Logout and revoke user token.",
                "operationId": "41de62ba059f25f3cc3255864519dff6",
                "responses": {
                    "200": {
                        "description": "Logout user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MiddlewareError"
                                },
                                "example": {
                                    "message": "User was logged out."
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Error: Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MiddlewareError"
                                },
                                "example": {
                                    "message": "Unauthenticated."
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "MiddlewareError": {
                "description": "Middleware error.",
                "properties": {
                    "message": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "Validation": {
                "description": "Shows validation errors.",
                "properties": {
                    "message": {
                        "type": "string"
                    },
                    "errors": {
                        "description": "Each key describes error message.",
                        "properties": {
                            "query or path name": {
                                "type": "array",
                                "items": {}
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object",
                "example": {
                    "message": "The email field must be a valid email address. (and 3 more errors)",
                    "errors": {
                        "email": [
                            "The email field must be a valid email address."
                        ],
                        "password": [
                            "The password field confirmation does not match.",
                            "The password field must contain at least one uppercase and one lowercase letter.",
                            "The password field must contain at least one number."
                        ]
                    }
                }
            },
...
            "Bearer": {
                "description": "Shows Bearer token",
                "properties": {
                    "accessToken": {
                        "description": "Bearer token",
                        "type": "string"
                    },
                    "token": {
                        "description": "Token info",
                        "properties": {
                            "id": {
                                "type": "string"
                            },
                            "user_id": {
                                "type": "integer"
                            },
                            "client_id": {
                                "type": "integer"
                            },
                            "name": {
                                "type": "string"
                            },
                            "scopes": {
                                "type": "array",
                                "items": {
                                    "properties": {}
                                }
                            },
                            "revoked": {
                                "type": "boolean"
                            },
                            "created_at": {
                                "type": "string"
                            },
                            "updated_at": {
                                "type": "string"
                            },
                            "expires_at": {
                                "type": "string"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "User": {
                "description": "Show information about the user",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "login": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "profilePhoto": {
                        "type": "string",
                        "nullable": true
                    },
                    "createdAt": {
                        "type": "string"
                    }
                },
                "type": "object",
                "example": {
                    "id": 3,
                    "login": "John_Doe",
                    "email": "john.doe@gmail.com",
                    "profilePhoto": null,
                    "createdAt": "2024-02-19 20:35:36"
                }
            }
        },
        "requestBodies": {
            "LoginRequest": {
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "email": {
                                    "type": "string"
                                },
                                "password": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        },
                        "example": {
                            "email": "",
                            "password": ""
                        }
                    }
                }
            },
            "UserRegisterRequest": {
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "login": {
                                    "description": "Create a new user.",
                                    "type": "string"
                                },
                                "email": {
                                    "type": "string"
                                },
                                "password": {
                                    "type": "string"
                                },
                                "password_confirmation": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        },
                        "example": {
                            "login": "",
                            "email": "",
                            "password": "",
                            "password_confirmation": ""
                        }
                    }
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}